Renamed view package to internal
