+++ app/src/androidTest/java/net/skyscanner/backpack/map/internal/BpkMapMarkerTest.kt
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package net.skyscanner.backpack.map.view
+package net.skyscanner.backpack.map.internal
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import net.skyscanner.backpack.BpkSnapshotTest
